Use git rev-parse for finding current branch
Summary: Git provides a `git rev-parse --abbrev-ref` command for
finding the current branch name directly. This is clearer to
understand than just looping through the human readable designed
branch listing. The previous implementation would fail if a branch
name had a `*` in the name. If the git repository is not on a branch
(in a detached head state) this will also not just fall back to
"master", but will instead return "HEAD" for the branch name.