[ci] [UW] Enable travis CI integration
This commit updates the legacy travis configuration file to run the
common tests with python3. It does not add client and cloud tests due
to packaging complexities. These will be added in a future commit.

It appears Travis was originally setup before the project was
incorporated by facebook, but the test infrastructure has rotted. This
commit updates build notifications to go to the private University of
Washington slack for now.