[smspdu] Account for DST when generating timestamps
The smspdu timestamp generation code uses python localtime(), which is
daylight savings aware. While it seems it would make the most sense to
actually always encode UTC or TAI and then display localtime on the
user device, it appears that the library decodes the user facing time
directly in the actual SMS message headers consistently. I will open
an issue to determine the right behavior.