[smspdu] Use naive byte <--> string mapping
The library uses strings to represent both unicode characters and
bytestrings. This is very unpythonic, but appears to be consistently
implemented throughout the library. It may make sense to update the
library to use bytes/bytearray when referring to bytes and string when
referring to unicode strings in the future for better built in error
checking/handling and to make the code more readable.